Implement dequeue mutex lock (ref. issue #1)
It would be a shame to leave behind this excellent project so I tried to fix issue #1 by implementing a mutex lock on dequeuing, so that no consumer could receive a pending task more than once (which was happening wiithout that). I tested that in my own application, and that measure seems to solve the problem.